@@ -16,29 +16,26 @@
     }
   ],
   "require": {
-    "php": "^7.4|^8.1",
-    "doctrine/annotations": "^1.11",
-    "symfony/framework-bundle": "^4.4|^5.4",
-    "symfony/event-dispatcher": "^4.4|^5.4",
-    "symfony/property-info": "^4.4|^5.4",
-    "symfony/validator": "^4.4|^5.4",
-    "symfony/property-access": "^4.4|^5.4",
-    "myclabs/php-enum": "^1.7",
-    "symfony/polyfill-php80": "^1"
+    "php": "^8.1",
+    "symfony/framework-bundle": "^5.4",
+    "symfony/event-dispatcher": "^5.4",
+    "symfony/property-info": "^5.4",
+    "symfony/validator": "^5.4",
+    "symfony/property-access": "^5.4"
   },
   "require-dev": {
     "ext-simplexml": "*",
     "phpunit/phpunit": "^9.5",
-    "psalm/plugin-symfony": "^2.4",
     "friendsofphp/php-cs-fixer": "^3",
     "pedrotroller/php-cs-custom-fixer": "^2",
-    "vimeo/psalm": "^4.10",
+    "vimeo/psalm": "^5",
     "phpstan/phpstan": "^1",
-    "symfony/phpunit-bridge": "^4.4|^5.3",
+    "symfony/phpunit-bridge": "^5.4",
     "matthiasnoback/symfony-dependency-injection-test": "^4.3",
     "nelmio/api-doc-bundle": "^4",
-    "symfony/stopwatch": "^4.4|^5.4",
-    "symfony/web-profiler-bundle": "^4.4|^5.4"
+    "symfony/stopwatch": "^5.4",
+    "symfony/web-profiler-bundle": "^5.4",
+    "doctrine/doctrine-bundle": "^2"
   },
   "conflict": {
     "nelmio/api-doc-bundle": "<4"
