@@ -33,7 +33,7 @@ class IntCoercer implements CoercerInterface
         }
 
         foreach ($value as $index => $val) {
-            if (is_numeric($val) && false === strpos((string)$val, '.')) {
+            if (is_numeric($val) && !str_contains((string)$val, '.')) {
                 $value[$index] = (int)$val;
             }
         }
