@@ -2,18 +2,14 @@
 
 namespace DM\DtoRequestBundle\Exception\DependencyInjection\Entity;
 
-class AnnotationMissingException extends \Exception
+class AttributeMissingException extends \Exception
 {
-    private string $class;
-
     public function __construct(
-        string $class,
+        private readonly string $class,
         string $message = "",
         int $code = 0,
         \Throwable $previous = null
     ) {
-        $this->class = $class;
-
         parent::__construct($message, $code, $previous);
     }
 
