@@ -5,6 +5,6 @@ namespace DM\DtoRequestBundle\Interfaces\Attribute;
 /**
  * Marker interface for knowing if an annotation is a dto one or not
  */
-interface DtoAnnotationInterface
+interface DtoAttributeInterface
 {
 }
