@@ -2,7 +2,7 @@
 
 namespace DM\DtoRequestBundle\Interfaces\Entity;
 
-use DM\DtoRequestBundle\Annotations\Dto\FindComplex;
+use DM\DtoRequestBundle\Attributes\Dto\FindComplex;
 
 /**
  * This interface must be implemented on services that later will be used with {@link FindComplex}
