@@ -2,12 +2,15 @@
 
 namespace DM\DtoRequestBundle\Profiler\Service\Resolver;
 
-use DM\DtoRequestBundle\Annotations\Dto\Bag;
+use DM\DtoRequestBundle\Attributes\Dto\Bag;
 use DM\DtoRequestBundle\Interfaces\Resolver\DtoTypeExtractorInterface;
 use DM\DtoRequestBundle\Model\Type\Dto;
 use DM\DtoRequestBundle\Profiler\AbstractWrapper;
 use Symfony\Component\Stopwatch\Stopwatch;
 
+/**
+ * @extends AbstractWrapper<Dto>
+ */
 class ProfilingDtoTypeExtractorService extends AbstractWrapper implements DtoTypeExtractorInterface
 {
     private DtoTypeExtractorInterface $dtoTypeExtractor;
