@@ -2,8 +2,8 @@
 
 namespace DM\DtoRequestBundle\Service\Resolver;
 
-use DM\DtoRequestBundle\Annotations\Dto\Bag;
-use DM\DtoRequestBundle\Annotations\Dto\Type as TypeAnnotation;
+use DM\DtoRequestBundle\Attributes\Dto\Bag;
+use DM\DtoRequestBundle\Attributes\Dto\Type as TypeAnnotation;
 use DM\DtoRequestBundle\Exception\Type\InvalidDateTimeClassException;
 use DM\DtoRequestBundle\Exception\Type\InvalidTypeCountException;
 use DM\DtoRequestBundle\Interfaces\Attribute\FindInterface;
@@ -11,21 +11,17 @@ use DM\DtoRequestBundle\Interfaces\DtoInterface;
 use DM\DtoRequestBundle\Interfaces\Resolver\DtoTypeExtractorInterface;
 use DM\DtoRequestBundle\Model\Type\Dto;
 use DM\DtoRequestBundle\Model\Type\Property;
-use Doctrine\Common\Annotations\Reader;
 use Symfony\Component\PropertyInfo\PropertyInfoExtractorInterface;
 use Symfony\Component\PropertyInfo\Type;
 
 class DtoTypeExtractorHelper implements DtoTypeExtractorInterface
 {
     private PropertyInfoExtractorInterface $propertyInfoExtractor;
-    private Reader $reader;
 
     public function __construct(
-        PropertyInfoExtractorInterface $propertyInfoExtractor,
-        Reader $reader
+        PropertyInfoExtractorInterface $propertyInfoExtractor
     ) {
         $this->propertyInfoExtractor = $propertyInfoExtractor;
-        $this->reader = $reader;
     }
 
     /**
@@ -44,24 +40,36 @@ class DtoTypeExtractorHelper implements DtoTypeExtractorInterface
         ?Bag $root = null
     ): Dto {
         $fqcn = $class->getName();
-        $root ??= $this->reader->getClassAnnotation($class, Bag::class) ?? new Bag();
+
+        if (null === $root) {
+            if (!empty($bag = $class->getAttributes(Bag::class))) {
+                $bag = $bag[0]->newInstance();
+            } else {
+                $bag = new Bag();
+            }
+
+            $root = $bag;
+        }
 
         $dto = new Dto();
 
-        foreach ($this->propertyInfoExtractor->getProperties($fqcn) as $property) {
+        foreach ($this->propertyInfoExtractor->getProperties($fqcn) ?? [] as $property) {
             if (!$this->propertyInfoExtractor->isWritable($fqcn, $property)) { // we won't be able to do anything with this anyway
                 continue;
             }
 
             try {
-                $annotations = $this->reader->getPropertyAnnotations(new \ReflectionProperty($fqcn, $property));
-            } catch (\ReflectionException $e) { // todo: php8 remove $e param, leave try-catch
+                $attributes = array_map(
+                    fn (\ReflectionAttribute $a) => $a->newInstance(),
+                    (new \ReflectionProperty($fqcn, $property))->getAttributes()
+                );
+            } catch (\ReflectionException) { // todo: php8 remove $e param, leave try-catch
                 continue;
             }
 
             $types = $this->propertyInfoExtractor->getTypes($fqcn, $property);
 
-            if (1 !== count($types)) {
+            if (1 !== count($types ?? [])) {
                 throw new InvalidTypeCountException(sprintf(
                     "Cannot deduct types with multiple specified types for property %s in class %s",
                     $property,
@@ -71,7 +79,7 @@ class DtoTypeExtractorHelper implements DtoTypeExtractorInterface
 
             $propertyClass = $this->getClass($types[0]);
 
-            if (is_subclass_of($propertyClass, DtoInterface::class)) {
+            if (null !== $propertyClass && is_subclass_of($propertyClass, DtoInterface::class)) {
                 /** @noinspection PhpUnhandledExceptionInspection */
                 $model = $this->extract(new \ReflectionClass($propertyClass), $root);
             } else {
@@ -79,47 +87,47 @@ class DtoTypeExtractorHelper implements DtoTypeExtractorInterface
             }
 
             // slightly special handling is required for this
-            /** @var FindInterface|null $findAnnotation */
-            $findAnnotation = array_values(array_filter($annotations, fn ($o) => $o instanceof FindInterface))[0] ?? null;
+            /** @var FindInterface|null $findAttribute */
+            $findAttribute = array_values(array_filter($attributes, fn ($o) => $o instanceof FindInterface))[0] ?? null;
 
             $model->setBag($root) // set default bag
                 ->setType($this->getType($types[0]) ?? 'string')
                 ->setCollection($types[0]->isCollection())
                 ->setParent($dto)
-                ->setPropertyAnnotations($annotations)
-                ->setFindAnnotation($findAnnotation)
+                ->setPropertyAttributes($attributes)
+                ->setFindAttribute($findAttribute)
                 ->setName($property)
                 ->setFqcn($propertyClass)
                 ->setDescription($this->propertyInfoExtractor->getShortDescription($fqcn, $property));
 
             $dto[$property] = $model;
 
-            if (null === $findAnnotation) {
+            if (null === $findAttribute) {
                 continue;
             }
 
-            foreach ($findAnnotation->getFields() as $key => $field) {
+            foreach ($findAttribute->getFields() as $key => $field) {
                 if (str_starts_with($field, '$')) { // dynamic
                     continue;
                 }
 
                 // simplify usage on the user end
-                if (!is_array($constraints = $findAnnotation->getConstraints()[$key] ?? [])) {
+                if (!is_array($constraints = $findAttribute->getConstraints()[$key] ?? [])) {
                     $constraints = [$constraints];
                 }
 
-                $type = $findAnnotation->getTypes()[$key] ?? new TypeAnnotation();
+                $type = $findAttribute->getTypes()[$key] ?? new TypeAnnotation();
 
                 $subProperty = (new Property())
                     ->setBag($model->getBag())
                     ->setParent($model)
                     ->setName($key)
-                    ->setPropertyAnnotations($constraints)
+                    ->setPropertyAttributes($constraints)
                     ->setType($type->type)
                     ->setSubType($type->subType)
                     ->setCollection($type->collection)
                     ->setFormat($type->format)
-                    ->setDescription($findAnnotation->getDescriptions()[$key] ?? null);
+                    ->setDescription($findAttribute->getDescriptions()[$key] ?? null);
 
                 /**
                  * @psalm-suppress InvalidArgument
@@ -136,13 +144,14 @@ class DtoTypeExtractorHelper implements DtoTypeExtractorInterface
     ): ?string {
         return !$type->isCollection() ?
             $type->getClassName() :
-            (null !== $type->getCollectionValueType() ? $type->getCollectionValueType()->getClassName() : null);
+            ($type->getCollectionValueTypes()[0] ?? null)?->getClassName() ?? null;
     }
 
     private function getType(
         Type $type
     ): ?string {
-        return !$type->isCollection() ? $type->getBuiltinType() :
-            (null !== $type->getCollectionValueType() ? $type->getCollectionValueType()->getBuiltinType() : null);
+        return !$type->isCollection() ?
+            $type->getBuiltinType() :
+            ($type->getCollectionValueTypes()[0] ?? null)?->getBuiltinType() ?? null;
     }
 }
