@@ -2,7 +2,7 @@
 
 namespace DM\DtoRequestBundle\Service\Type\Coercer;
 
-use DM\DtoRequestBundle\Annotations\Dto\Format;
+use DM\DtoRequestBundle\Attributes\Dto\Format;
 use DM\DtoRequestBundle\Interfaces\Type\CoercerInterface;
 use DM\DtoRequestBundle\Model\Type\CoerceResult;
 use DM\DtoRequestBundle\Model\Type\Property;
@@ -57,7 +57,7 @@ class DateTimeImmutableCoercer implements CoercerInterface
                 $value[$index] = $time;
             }
         } else {
-            if (null === $value || false === ($time = \DateTimeImmutable::createFromFormat($format, $value))) {
+            if (null === $value || false === ($time = \DateTimeImmutable::createFromFormat($format, $value))) { // @phpstan-ignore-line
                 $value = null;
             } else {
                 $value = $time;
