@@ -2,20 +2,19 @@
 
 namespace DM\DtoRequestBundle\Service\Type\Coercer;
 
-use DM\DtoRequestBundle\Annotations\Dto\FromKey;
+use DM\DtoRequestBundle\Attributes\Dto\FromKey;
 use DM\DtoRequestBundle\Interfaces\Type\CoercerInterface;
 use DM\DtoRequestBundle\Model\Type\CoerceResult;
 use DM\DtoRequestBundle\Model\Type\Property;
 use DM\DtoRequestBundle\Traits\Type\CoerceConstructWithValidatorTrait;
 use DM\DtoRequestBundle\Traits\Type\CoercerResultTrait;
 use DM\DtoRequestBundle\Util;
-use MyCLabs\Enum\Enum;
 use Symfony\Component\Validator\Constraints\All;
 use Symfony\Component\Validator\Constraints\Choice;
 use Symfony\Component\Validator\ConstraintViolationInterface;
 
 /**
- * @implements CoercerInterface<Enum|null>
+ * @implements CoercerInterface<\BackedEnum|null>
  */
 class EnumCoercer implements CoercerInterface
 {
@@ -23,12 +22,17 @@ class EnumCoercer implements CoercerInterface
     use CoerceConstructWithValidatorTrait;
 
     private bool $fromKey = false;
-    private string $class;
+
+    /**
+     * @var list<\BackedEnum>
+     */
+    private array $cases;
 
     public function supports(
         Property $property
     ): bool {
-        return 'object' === $property->getType() && is_subclass_of($property->getFqcn(), Enum::class);
+        return 'object' === $property->getType() &&
+            is_subclass_of($property->getFqcn() ?? '', \BackedEnum::class);
     }
 
     public function coerce(
@@ -36,8 +40,9 @@ class EnumCoercer implements CoercerInterface
         Property $property,
         $value
     ): CoerceResult {
-        $this->fromKey = $property->hasDtoAnnotation(FromKey::class);
-        $this->class = $property->getFqcn();
+        $this->fromKey = $property->hasDtoAttribute(FromKey::class);
+
+        $this->cases = call_user_func([$property->getFqcn(), 'cases']); // @phpstan-ignore-line
         $constraints = [new Choice(['choices' => $property->getEnumChoices()])];
 
         if ($property->isCollection()) {
@@ -55,7 +60,7 @@ class EnumCoercer implements CoercerInterface
 
         if (!$property->isCollection()) {
             return new CoerceResult(
-                0 === $violations->count() ? $this->createEnum($value) : null,
+                0 === $violations->count() ? $this->createEnum($value) : null, // @phpstan-ignore-line
                 $violations
             );
         }
@@ -75,24 +80,21 @@ class EnumCoercer implements CoercerInterface
         );
     }
 
-    /**
-     * @param int|string|null $value
-     *
-     * @return Enum|null
-     */
     private function createEnum(
-        $value
-    ): ?Enum {
+        int|string|null $value
+    ): \BackedEnum|null {
         if (null === $value) {
             return null;
         }
 
-        if ($this->fromKey) {
-            // @phpstan-ignore-next-line
-            return call_user_func([$this->class, $value]);
+        foreach ($this->cases as $case) {
+            $compare = $this->fromKey ? $case->name : $case->value;
+
+            if ($value === $compare) {
+                return $case;
+            }
         }
 
-        // @phpstan-ignore-next-line
-        return call_user_func([$this->class, 'from'], $value);
+        return null;
     }
 }
