@@ -4,14 +4,20 @@ namespace DM\DtoRequestBundle\Traits\Annotation;
 
 trait HttpActionTrait
 {
-    public int $statusCode;
-    public ?string $message;
-    public ?string $description;
-
     /**
-     * @var array<string, string>
+     * @param int $statusCode
+     * @param string|null $message
+     * @param array $headers
+     * @psalm-param array<string, string> $headers
+     * @param string|null $description
      */
-    public array $headers;
+    public function __construct(
+        public int $statusCode,
+        public string|null $message = '',
+        public array $headers = [],
+        public string|null $description = null
+    ) {
+    }
 
     public function getHttpStatusCode(): int
     {
