@@ -2,15 +2,9 @@
 
 namespace DM\DtoRequestBundle\Tests\Fixtures\Enum;
 
-use DM\DtoRequestBundle\Interfaces\Enum\IntegerBackedEnumInterface;
-use MyCLabs\Enum\Enum;
-
-/**
- * @psalm-immutable
- */
-class IntegerEnum extends Enum implements IntegerBackedEnumInterface
+enum IntegerEnum: int
 {
-    public const INTEGER_KEY = 15;
-    public const OTHER_KEY = 20;
-    public const LAST_KEY = 25;
+    case INTEGER_KEY = 15;
+    case OTHER_KEY = 20;
+    case LAST_KEY = 25;
 }
