@@ -2,13 +2,12 @@
 
 namespace DM\DtoRequestBundle\Tests\Fixtures\Model\Dto;
 
-use DM\DtoRequestBundle\Annotations\Dto\Bag;
+use DM\DtoRequestBundle\Attributes\Dto\Bag;
+use DM\DtoRequestBundle\Enum\BagEnum;
 use DM\DtoRequestBundle\Model\AbstractDto;
 use DM\DtoRequestBundle\Tests\Fixtures\Enum\IntegerEnum;
 
-/**
- * @Bag("query")
- */
+#[Bag(BagEnum::Query)]
 class EnumQueryDto extends AbstractDto
 {
     public ?IntegerEnum $int = null;
