@@ -2,20 +2,18 @@
 
 namespace DM\DtoRequestBundle\Tests\Fixtures\Model\Dto;
 
-use DM\DtoRequestBundle\Annotations\Dto\FindOneBy;
-use DM\DtoRequestBundle\Annotations\Dto\Type;
+use DM\DtoRequestBundle\Attributes\Dto\FindOneBy;
+use DM\DtoRequestBundle\Attributes\Dto\Type;
 use DM\DtoRequestBundle\Model\AbstractDto;
 use DM\DtoRequestBundle\Tests\Fixtures\Model\DummyModel;
 use Symfony\Component\Validator\Constraints as Assert;
 
 class FindWithSomeSecondErrorDto extends AbstractDto
 {
-    /**
-     * @FindOneBy(
-     *     fields={"id": "something_id", "second": "something_second"},
-     *     types={"id": @Type("int")},
-     *     constraints={"id": @Assert\NotBlank(), "second": @Assert\NotBlank()}
-     * )
-     */
+    #[FindOneBy(
+        fields: ['id' => 'something_id', 'second' => 'something_second'],
+        constraints: ['id' => new Assert\NotBlank(), 'second' => new Assert\NotBlank()],
+        types: ['id' => new Type('int')]
+    )]
     public ?DummyModel $model = null;
 }
