@@ -2,16 +2,15 @@
 
 namespace DM\DtoRequestBundle\Tests\Fixtures\Model\HttpAction;
 
-use DM\DtoRequestBundle\Annotations\Dto\FindOneBy;
-use DM\DtoRequestBundle\Annotations\Dto\Http\OnNull;
+use DM\DtoRequestBundle\Attributes\Dto\FindOneBy;
+use DM\DtoRequestBundle\Attributes\Dto\Http\OnNull;
 use DM\DtoRequestBundle\Model\AbstractDto;
 use DM\DtoRequestBundle\Tests\Fixtures\Model\DummyModel;
+use Symfony\Component\HttpFoundation\Response;
 
 class FindOneOnNullDto extends AbstractDto
 {
-    /**
-     * @OnNull(404)
-     * @FindOneBy(fields={"id": "something"})
-     */
+    #[OnNull(Response::HTTP_NOT_FOUND)]
+    #[FindOneBy(fields: ['id' => 'something'])]
     public ?DummyModel $model = null;
 }
