@@ -2,14 +2,14 @@
 
 namespace DM\DtoRequestBundle\Tests\Fixtures\Model\PathFixDto;
 
-use DM\DtoRequestBundle\Annotations\Dto\Path;
+use DM\DtoRequestBundle\Attributes\Dto\Path;
 use DM\DtoRequestBundle\Model\AbstractDto;
 
 class RequestEdgeCaseDto extends AbstractDto
 {
     /**
-     * @Path("")
      * @var PathFixDto[]
      */
+    #[Path('')]
     public array $dtos = [];
 }
