@@ -2,15 +2,16 @@
 
 namespace DM\DtoRequestBundle\Tests\Service\Entity;
 
-use DM\DtoRequestBundle\Annotations\Entity\EntityProvider;
+use DM\DtoRequestBundle\Attributes\Entity\EntityProvider;
 use DM\DtoRequestBundle\Interfaces\Entity\ProviderInterface;
 use DM\DtoRequestBundle\Tests\Fixtures\Model\DummyModel;
 
 /**
- * @EntityProvider(DummyModel::class, true)
- *
  * @see DummyModel
+ *
+ * @implements ProviderInterface<DummyModel>
  */
+#[EntityProvider(DummyModel::class, true)]
 class DummyModelProvider implements ProviderInterface
 {
     public function findComplex(
